+++ src/main/webapp/dao/TraineeHibernateDao.java
@@ -17,6 +17,6 @@ import pojo.Trainee;
  * 
  * @author root
  */
-public class TraineeHibernateDao extends GenericHibernateDAO<Trainee, Long> implements TraineeDao{
+public class TraineeHibernateDao extends GenericHibernateDAO<Trainee, Integer> implements TraineeDao{
     
 }
